[script] add example file move script, refactored script duplicated code.
