+++ .gitignore
@@ -110,3 +110,4 @@ dist
 build
 .aptos
 *.log
+*.mv
