+++ scripts/deploy.ts
@@ -1,13 +1,9 @@
 import {DEF_ACCOUNT_CONF, MY_ACCOUNT} from "../src/web3/global";
 import {Command} from "commander";
 import {HexString} from "aptos";
-import {isHexEqual, isStringAddress} from "../src/utils/check";
 import {MovePublisher} from "../src/momentum-safe/move-publisher";
-import {MomentumSafe} from "../src/momentum-safe/momentum-safe";
-import * as Aptos from "../src/web3/global";
 import {makeModulePublishTx} from "../src/momentum-safe/msafe-txn";
-import {printSeparator, printTxDetails, promptForYN} from "../src/cmd/common";
-import {loadMomentumSafe} from "./common";
+import {loadMomentumSafe, parseAndLoadTxnConfig, printTxnAndConfirm, proposeTransaction} from "./common";
 import {DEFAULT_ENDPOINT, DEFAULT_FAUCET, DEFAULT_MSAFE, DEFAULT_NETWORK, loadConfigAndApply} from "../src/utils/load";
 
 const program = new Command();
@@ -31,16 +27,17 @@ const cli = program
 
 
 async function main() {
-  const args = await parseAndLoadConfig();
+  const args = await parseAndLoadTxnConfig(cli);
+  const moveDir = getMoveDir(cli);
 
   // load msafe
-  const msafe = await loadMomentumSafe(HexString.ensure(args.msafeAddr));
+  const msafe = await loadMomentumSafe(HexString.ensure(args.msafe));
 
   // make module publish transaction
   const sn = await msafe.getNextSN();
   const msafeTxn = await makeModulePublishTx(
     msafe,
-    {moveDir: args.moveDir},
+    {moveDir: moveDir},
     {
       sequenceNumber: sn,
       gasPrice: args.gasPrice,
@@ -51,80 +48,20 @@ async function main() {
   );
 
   // Confirm transaction details
-  await printTxDetails(msafeTxn.getTxnInfo());
-  printSeparator();
-  const userConfirm = await promptForYN("Do you confirm with the transaction?", true);
-  if (!userConfirm) {
-    console.error("User canceled operation");
-    process.exit(1);
-  }
-
+  await printTxnAndConfirm(msafeTxn);
   // Submit transaction
-  const res = await msafe.initTransaction(MY_ACCOUNT, msafeTxn, {
-    gasPrice: args.gasPrice,
-    maxGas: args.maxGas,
-    estimateMaxGas: args.estimateMaxGas,
-    estimateGasPrice: args.estimateGasPrice,
-  });
-  const myHash = (res.pendingTx as any).hash;
-  console.log(`\tTransaction ${myHash} submitted to blockchain`);
-  await Aptos.waitForTransaction(myHash);
-  console.log(`\tTransaction confirmed on chain.`);
+  await proposeTransaction(msafe, msafeTxn, args);
 }
 
-async function parseAndLoadConfig(): Promise<configArg> {
-  const args = getArguments();
-  validateArguments(args);
-
-  await loadConfigAndApply({
-    configFilePath: args.config,
-    profile: args.profile,
-    network: args.network,
-    endpoint: args.endpoint,
-    faucet: args.faucet,
-    msafeDeployer: args.msafeDeployer,
-  });
-  return args;
+function getMoveDir(cli: Command) {
+  const moveDir = cli.opts().moveDir;
+  validateMoveDir(moveDir);
+  return moveDir;
 }
 
-type configArg = {
-  config: string,
-  profile: string,
-  network: string,
-  moveDir: string,
-  maxGas: bigint,
-  estimateMaxGas: boolean,
-  gasPrice: bigint,
-  estimateGasPrice: boolean,
-  endpoint: string,
-  faucet: string,
-  msafeDeployer: string,
-  msafeAddr: string,
-}
-
-function getArguments(): configArg {
-  const estimateGasPrice = cli.opts().gasPrice === undefined;
-  const estimateMaxGas = cli.opts().maxGas === undefined;
-
-  return {
-    config: cli.opts().config,
-    profile: cli.opts().profile,
-    network: cli.opts().network.toLowerCase(),
-    moveDir: cli.opts().moveDir,
-    maxGas: cli.opts().maxGas,
-    gasPrice: cli.opts().gasPrice,
-    estimateGasPrice,
-    estimateMaxGas,
-    endpoint: cli.opts().endpoint.toLowerCase(),
-    faucet: cli.opts().faucet.toLowerCase(),
-    msafeDeployer: cli.opts().msafeDeployer.toLowerCase(),
-    msafeAddr: cli.opts().msafe.toLowerCase(),
-  };
-}
-
-function validateArguments(ca: configArg) {
-  if (!MovePublisher.isDirValid(ca.moveDir)) {
-    throw Error("invalid move dir: " + ca.msafeDeployer);
+function validateMoveDir(moveDir: string) {
+  if (!MovePublisher.isDirValid(moveDir)) {
+    throw Error("invalid move dir: " + moveDir);
   }
 }
