+++ tests/script/sources/test_script.move
@@ -1,0 +1,7 @@
+script {
+    use aptos_framework::coin;
+
+    fun main<T>(s: &signer, to: address, amount: u64) {
+        coin::transfer<T>(s, to, amount);
+    }
+}
